health ai agent/requirements.txt
google-genai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
rich>=13.7.0
typer>=0.12.0
python-dateutil>=2.9.0
